refactor: move filterByTag logic from activation.ts to commands.ts
